tctl resources: convert windows_desktop_service to new handler
Updates #60370
